+++ tests/test_application.py
@@ -1,14 +1,9 @@
-from unittest.mock import Mock
-
 import os
 
-from jupyterhub import orm
 from tests.temp_mixin import TempMixin
-from tornado import gen, testing
+from tornado import testing
 
 from remoteappmanager.application import Application
-from remoteappmanager.docker.container import Container
-from remoteappmanager.docker.container_manager import ContainerManager
 from tests import utils
 
 
@@ -42,45 +37,8 @@ class TestApplication(TempMixin, testing.AsyncTestCase):
         self.assertIsNotNone(app.command_line_config)
         self.assertIsNotNone(app.file_config)
 
-        # Test the configuration options
-        self.assertIsNotNone(app.command_line_config.port)
-        self.assertIsInstance(app.container_manager, ContainerManager)
-        self.assertIsInstance(app.reverse_proxy, orm.Proxy)
-
-    def test_container_url_abspath(self):
-        app = self.app
-        container = Container(url_id="12345")
-        abspath = app.container_url_abspath(container)
-        self.assertEqual(abspath, "/user/username/containers/12345")
-
-    @testing.gen_test
-    def test_reverse_proxy_operations(self):
-        coroutine_out = None
-
-        @gen.coroutine
-        def mock_api_request(self, *args, **kwargs):
-            nonlocal coroutine_out
-            yield gen.sleep(0.1)
-            coroutine_out = dict(args=args, kwargs=kwargs)
-
-        app = self.app
-        app.reverse_proxy = Mock(spec=orm.Proxy)
-        app.reverse_proxy.api_request = mock_api_request
-
-        container = Container(docker_id="12345")
-        yield app.reverse_proxy_add_container(container)
-
-        self.assertEqual(coroutine_out["kwargs"]["method"], "POST")
-
-        yield app.reverse_proxy_remove_container(container)
-
-        self.assertEqual(coroutine_out["kwargs"]["method"], "DELETE")
-
-    def test_database_initialization(self):
-        app = self.app
-
         self.assertIsNotNone(app.db)
         self.assertIsNotNone(app.user)
-
-        self.assertEqual(app.user.name, "username")
-        self.assertEqual(app.user.orm_user, None)
+        self.assertIsNotNone(app.reverse_proxy)
+        self.assertIsNotNone(app.container_manager)
+        self.assertIsNotNone(app.hub)
